fix showing snackbars when not logged in
